+++ src/driving_distance/src/boost_interface_drivedist.h
@@ -32,8 +32,8 @@ extern "C" {
 #endif
 
 void
-do_pgr_driving_many_to_dist(pgr_edge_t  * edges, int64_t total_tuples,
-                      int64_t  *start_vertex, int s_len,
+do_pgr_driving_many_to_dist(pgr_edge_t  * edges, size_t total_tuples,
+                      int64_t  *start_vertex, size_t s_len,
                       float8 distance,
                       bool directed,
                       bool equicost,
@@ -41,7 +41,7 @@ do_pgr_driving_many_to_dist(pgr_edge_t  * edges, int64_t total_tuples,
                       char ** err_msg);
 
 void
-do_pgr_driving_distance(pgr_edge_t  * edges, int64_t total_tuples,
+do_pgr_driving_distance(pgr_edge_t  * edges, size_t total_tuples,
                       int64_t  start_vertex, float8 distance,
                       bool directed,
                       General_path_element_t **ret_path, size_t *path_count,
